refactor(lowlevel): move envelope logic to separate package
