+++ internal/lowlevel/envelope/mod.go
@@ -1,4 +1,4 @@
-package lowlevel
+package envelope
 
 import (
 	"crypto/rand"
