+++ internal/lowlevel/envelope/mod_test.go
@@ -1,4 +1,4 @@
-package lowlevel
+package envelope
 
 import (
 	"encoding/xml"
